医生和导师保存时出现NAN的问题

+++ src/main/java/com/xmheart/service/impl/TeacherTeamServiceImpl.java
@@ -121,7 +121,7 @@ public class TeacherTeamServiceImpl implements TeacherTeamService {
 //            return -1;
 //        }
         
-        Byte order1 = doctor1.getDocOrder();
+        Integer order1 = doctor1.getDocOrder();
         doctor1.setDocOrder(doctor2.getDocOrder());
         xpwTeacherMapper.updateByPrimaryKey(doctor1);
